requirements.txt
asgiref==3.9.1
Django==5.2.6
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
ordered-set==4.1.0
prettytable==3.16.0
PyJWT==2.10.1
pymongo==4.13.2
PyMySQL==1.1.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
wcwidth==0.2.13
